+++ realm/realm-library/src/main/java/io/realm/internal/async/QueryUpdateTask.java
@@ -31,7 +31,7 @@ import io.realm.internal.RealmObjectProxy;
 import io.realm.internal.SharedRealm;
 import io.realm.internal.Table;
 import io.realm.internal.TableQuery;
-import io.realm.internal.log.RealmLog;
+import io.realm.log.RealmLog;
 
 /**
  * Manages the update of async queries.
@@ -100,11 +100,11 @@ public class QueryUpdateTask implements Runnable {
 
         } catch (BadVersionException e) {
             // In some rare race conditions, this can happen. In that case, just ignore the error.
-            RealmLog.d("Query update task could not complete due to a BadVersionException. " +
+            RealmLog.debug("Query update task could not complete due to a BadVersionException. " +
                     "Retry is scheduled by a REALM_CHANGED event.");
 
         } catch (Throwable e) {
-            RealmLog.e(e.getMessage(), e);
+            RealmLog.error(e);
             Handler handler = callerHandler.get();
             if (handler != null && handler.getLooper().getThread().isAlive()) {
                 handler.obtainMessage(HandlerControllerConstants.REALM_ASYNC_BACKGROUND_EXCEPTION, new Error(e)).sendToTarget();